+++ source/benoit/benoit.rs
@@ -36,7 +36,7 @@ pub mod video;
 pub const VERSION: (u32, u32, u32) = (
 	0x2, // Major
 	0x6, // Minor
-	0x1, // Patch
+	0x2, // Patch
 );
 
 pub const PRECISION: u32 = 0x80;